requirements.txt
langgraph>=0.2.0
langgraph-checkpoint-sqlite>=1.0.0
openai>=1.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
sentence-transformers>=3.0.0
numpy>=1.26.0
pytest>=8.0.0
langfuse>=4.0.0
